fix(spec): Refine formal analysis and (D)TLS language
- Updated language in reference to work-in-progress formal analysis
(ProVerif) to be precise regarding the maturity of the analysis in
progress. E.g., using  "preliminary," "ongoing," or "being modelled"
instead of "formally verified." This affects the abstract, security
considerations, and Appendix A.

- The abstract is updated to clarify that the protocol *utilizes* the
attested TLS draft [draft-fossati-tls-exported-attestation] to *define*
its attestation renewal procedure, not that it defines the attested TLS
protocol itself. This correctly scopes the draft's relationship to
attested (D)TLS.